requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.4.0
python-multipart==0.0.6
pydantic==2.4.0
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
